+++ spec/immer/store.spec.ts
@@ -91,23 +91,23 @@ describe('Store tests - Immer', () => {
             expect(StateKeeper.CURRENT_STATE['locallyInitialized']['nestedValue']['value']).toEqual('v1');
         });
 
-        it('should thorw exception on try to reset state when state path points to value but not an object', () => {
+        it('should throw exception on try to reset state when state path points to value but not an object', () => {
             expect(() => store.select(['layout', 'test']).reset()).toThrowError();
         });
 
         it('should reset state when store has initial state', () => {
-            const intilizedStore = store.initialize(['actionStore'], { test: { url: 'home' } });
-            intilizedStore.select(['test']).update(state => state['url'] = 'home-updated');
+            const initializedStore = store.initialize(['actionStore'], { test: { url: 'home' } });
+            initializedStore.select(['test']).update(state => state['url'] = 'home-updated');
             expect(StateKeeper.CURRENT_STATE['actionStore']['test']['url']).toEqual('home-updated');
 
-            intilizedStore.select(['test']).reset();
+            initializedStore.select(['test']).reset();
             expect(StateKeeper.CURRENT_STATE['actionStore']['test']['url']).toEqual('home');
         });
 
         it('should rootPath and initialState vallues to store after initialization', () => {
-            const intilizedStore = store.initialize(['actionStore'], { test: { url: 'home' } });
-            expect(intilizedStore.rootPath).toContain('actionStore');
-            expect(intilizedStore.initialState.test.url).toBe('home');
+            const initializedStore = store.initialize(['actionStore'], { test: { url: 'home' } });
+            expect(initializedStore.rootPath).toContain('actionStore');
+            expect(initializedStore.initialState.test.url).toBe('home');
         });
     });
 });